+++ app/not-found.tsx
@@ -1,14 +1,14 @@
-import Link from 'next/link'
 import { Button } from 'flowbite-react'
+import Link from 'next/link'
 
 export default function NotFound() {
   return (
     <div className="flex flex-col items-center gap-4 h-full mt-8">
       <h2 className="text-5xl font-semibold">Not Found</h2>
       <p>Could not find requested resource</p>
-      <Button as={Link} prefetch={false} href="/" color="gray">
-        Go to Home
-      </Button>
+      <Link href="/">
+        <Button color="gray">Go to Home</Button>
+      </Link>
     </div>
   )
 }
