+++ middlewares/verifySession.ts
@@ -1,9 +1,9 @@
 import { getServerSession } from 'next-auth'
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import authOptions from '@/utils/authOptions'
 import { ROLE } from '@prisma/client'
 import { NextResponse } from 'next/server'
 import { ENV, STATUS_CODE } from '@/utils/constant'
-import StandardResponse from '@/utils/types/StandardResponse'
+import standardResponse from '@/utils/standardResponese'
 
 const verifySession = async (req: any, next: any) => {
   if (ENV.NODE_ENV === 'development') {
@@ -13,25 +13,15 @@ const verifySession = async (req: any, next: any) => {
   const session = await getServerSession(authOptions)
 
   if (!session || !session.user) {
-    return NextResponse.json(
-      {
-        statusCode: STATUS_CODE.UNAUTHORIZED,
-        message: 'Unauthorized',
-        data: null,
-      } satisfies StandardResponse,
-      { status: STATUS_CODE.UNAUTHORIZED },
-    )
+    return NextResponse.json(standardResponse(STATUS_CODE.UNAUTHORIZED, 'Unauthorized', null), {
+      status: STATUS_CODE.UNAUTHORIZED,
+    })
   }
 
   if (session.user.role !== ROLE.ADMIN) {
-    return NextResponse.json(
-      {
-        statusCode: STATUS_CODE.FORBIDDEN,
-        message: 'Forbidden',
-        data: null,
-      } satisfies StandardResponse,
-      { status: STATUS_CODE.FORBIDDEN },
-    )
+    return NextResponse.json(standardResponse(STATUS_CODE.FORBIDDEN, 'Forbidden', null), {
+      status: STATUS_CODE.FORBIDDEN,
+    })
   }
 
   next()
